Chore: Move sku queries to inside the sku component
