+++ react/components/SKUTranslation/SKUForm.tsx
@@ -5,7 +5,7 @@ import { Input } from 'vtex.styleguide'
 import { useAlert } from '../../providers/AlertProvider'
 import useFormTranslation from '../../hooks/useFormTranslation'
 import { useLocaleSelector } from '../LocaleSelector'
-import translateSKUMutation from '../../graphql/translateSKU.gql'
+import translateSKUMutation from './graphql/translateSKU.gql'
 import ActionButtons from '../ActionButtons'
 
 interface SKUFormProps {
