+++ react/components/SKUTranslation/graphql/downloadSkuTranslations.gql
