+++ react/components/SKUTranslation/graphql/startSkuTranslations.gql
