+++ react/components/SKUTranslation/graphql/translateSKU.gql
